requirements.txt
# Web框架
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
chainlit>=1.0.0
python-multipart>=0.0.6

# Langchain框架
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0

# 文档解析
python-docx>=0.8.11
PyPDF2>=3.0.0
python-pptx>=0.6.21

# 向量数据库
chromadb>=0.4.0

# 其他工具
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
openai>=1.0.0

# 嵌入模型（使用sentence-transformers替代HuggingFaceEmbeddings）
sentence-transformers>=2.2.0
